Rename MultisetPropertyFactory.CodeGenerator -> MultisetProperty
MultisetPropertyFactory is now MultisetProperty.Factory
